fix: Resolve CI type-checking failures and improve type safety across codebase
- Replace bare type: ignore comments with specific error codes (e.g.,
  no-any-return, assignment, attr-defined) for precise suppression
- Add unused-ignore safety net to all type: ignore comments for
  cross-environment compatibility (local vs CI mypy configurations)
- Wrap untyped return values with np.asarray(), float(), int(), bool(),
  and typing.cast() to improve actual type safety
- Fix torch.stack type: ignore placement in adaptive_fusion.py
  (must be on opening line for mypy to associate correctly)
- Restore --ignore-missing-imports flag to CI mypy invocation
- Add torch model cache to CI for faster ML test runs
- Switch opencv-python to opencv-python-headless for CI compatibility

+++ src/omni_mercury_engine/biometric/iris_recognition.py
@@ -712,6 +712,10 @@ class IrisRecognizer:
 
         contrast = np.std(normalized_iris[mask])
 
-        quality = 0.4 * usable_ratio + 0.3 * min(1.0, sharpness * 10) + 0.3 * min(1.0, contrast * 5)  # type: ignore[operator, unused-ignore]
+        quality = (  # type: ignore[operator, unused-ignore]
+            0.4 * usable_ratio
+            + 0.3 * min(1.0, float(sharpness * 10))
+            + 0.3 * min(1.0, float(contrast * 5))
+        )
 
         return float(quality)
